+++ packages/services/guard/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @0xsequence/guard
 
+## 3.0.10
+
+### Patch Changes
+
+- Minor network updates, relayer interface
+
+## 3.0.9
+
+### Patch Changes
+
+- Fee options fixes
+
+## 3.0.8
+
+### Patch Changes
+
+- Bug fix for relayer fee options handling
+
 ## 3.0.7
 
 ### Patch Changes